LPS-106413 Moves to layout.type.controller package
