+++ modules/apps/layout/layout-type-controller/layout-type-controller-display-page/src/main/java/com/liferay/layout/type/controller/display/page/internal/layout/type/controller/DisplayPageLayoutTypeController.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.content.internal.controller;
+package com.liferay.layout.type.controller.display.page.internal.layout.type.controller;
 
 import com.liferay.fragment.constants.FragmentActionKeys;
 import com.liferay.fragment.contributor.FragmentCollectionContributorTracker;
@@ -20,15 +20,14 @@ import com.liferay.fragment.renderer.FragmentRendererController;
 import com.liferay.fragment.renderer.FragmentRendererTracker;
 import com.liferay.info.constants.InfoDisplayWebKeys;
 import com.liferay.info.display.contributor.InfoDisplayContributorTracker;
+import com.liferay.info.display.request.attributes.contributor.InfoDisplayRequestAttributesContributor;
 import com.liferay.info.item.renderer.InfoItemRendererTracker;
 import com.liferay.info.item.selector.InfoItemSelectorTracker;
 import com.liferay.item.selector.ItemSelector;
 import com.liferay.layout.content.page.editor.constants.ContentPageEditorWebKeys;
 import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
 import com.liferay.layout.page.template.service.LayoutPageTemplateEntryLocalService;
-import com.liferay.layout.security.permission.resource.LayoutContentModelResourcePermission;
 import com.liferay.layout.type.controller.BaseLayoutTypeControllerImpl;
-import com.liferay.layout.type.controller.content.internal.constants.ContentLayoutTypeControllerWebKeys;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
@@ -41,7 +40,7 @@ import com.liferay.portal.kernel.security.auth.PrincipalException;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
 import com.liferay.portal.kernel.service.LayoutLocalService;
-import com.liferay.portal.kernel.service.permission.LayoutPermission;
+import com.liferay.portal.kernel.service.permission.LayoutPermissionUtil;
 import com.liferay.portal.kernel.servlet.TransferHeadersHelperUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.Constants;
@@ -49,6 +48,8 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.taglib.servlet.PipingServletResponse;
+
+import java.util.List;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
@@ -63,10 +64,12 @@ import org.osgi.service.component.annotations.Reference;
  * @author Juergen Kappler
  */
 @Component(
-	immediate = true, property = "layout.type=" + LayoutConstants.TYPE_CONTENT,
+	immediate = true,
+	property = "layout.type=" + LayoutConstants.TYPE_ASSET_DISPLAY,
 	service = LayoutTypeController.class
 )
-public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
+public class DisplayPageLayoutTypeController
+	extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public String getType() {
@@ -121,28 +124,22 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 		if (layoutMode.equals(Constants.EDIT)) {
 			httpServletRequest.setAttribute(
-				ContentLayoutTypeControllerWebKeys.ITEM_SELECTOR,
-				_itemSelector);
-			httpServletRequest.setAttribute(
 				ContentPageEditorWebKeys.
 					FRAGMENT_COLLECTION_CONTRIBUTOR_TRACKER,
 				_fragmentCollectionContributorTracker);
+
 			httpServletRequest.setAttribute(
-				FragmentActionKeys.FRAGMENT_RENDERER_TRACKER,
-				_fragmentRendererTracker);
+				ContentPageEditorWebKeys.ITEM_SELECTOR, _itemSelector);
 		}
 
 		httpServletRequest.setAttribute(
 			FragmentActionKeys.FRAGMENT_RENDERER_CONTROLLER,
 			_fragmentRendererController);
-		httpServletRequest.setAttribute(
-			InfoDisplayWebKeys.INFO_DISPLAY_CONTRIBUTOR_TRACKER,
-			_infoDisplayContributorTracker);
 
 		String page = getViewPage();
 
 		if (layoutMode.equals(Constants.EDIT)) {
-			page = _EDIT_LAYOUT_PAGE;
+			page = _EDIT_PAGE;
 		}
 
 		RequestDispatcher requestDispatcher =
@@ -160,11 +157,8 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 			RequestDispatcher.INCLUDE_SERVLET_PATH);
 
 		try {
-			LayoutPageTemplateEntry layoutPageTemplateEntry = null;
-
-			if (layoutMode.equals(Constants.EDIT)) {
-				layoutPageTemplateEntry = _fetchLayoutPageTemplateEntry(layout);
-			}
+			LayoutPageTemplateEntry layoutPageTemplateEntry =
+				_fetchLayoutPageTemplateEntry(layout);
 
 			if (layoutPageTemplateEntry != null) {
 				httpServletRequest.setAttribute(
@@ -175,14 +169,13 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 					ContentPageEditorWebKeys.CLASS_PK,
 					layoutPageTemplateEntry.getLayoutPageTemplateEntryId());
 			}
-			else {
-				httpServletRequest.setAttribute(
-					ContentPageEditorWebKeys.CLASS_NAME,
-					Layout.class.getName());
 
-				httpServletRequest.setAttribute(
-					ContentPageEditorWebKeys.CLASS_PK, layout.getPlid());
-			}
+			httpServletRequest.setAttribute(
+				FragmentActionKeys.FRAGMENT_RENDERER_TRACKER,
+				_fragmentRendererTracker);
+			httpServletRequest.setAttribute(
+				InfoDisplayWebKeys.INFO_DISPLAY_CONTRIBUTOR_TRACKER,
+				_infoDisplayContributorTracker);
 
 			addAttributes(httpServletRequest);
 
@@ -206,38 +199,44 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 	}
 
 	@Override
-	public boolean isBrowsable() {
-		return true;
-	}
-
-	@Override
 	public boolean isFirstPageable() {
 		return true;
 	}
 
 	@Override
 	public boolean isFullPageDisplayable() {
+		return true;
+	}
+
+	@Override
+	public boolean isInstanceable() {
 		return false;
 	}
 
 	@Override
 	public boolean isParentable() {
-		return true;
-	}
-
-	@Override
-	public boolean isPrimaryType() {
-		return true;
+		return false;
 	}
 
 	@Override
 	public boolean isSitemapable() {
-		return true;
+		return false;
 	}
 
 	@Override
 	public boolean isURLFriendliable() {
 		return true;
+	}
+
+	@Override
+	protected void addAttributes(HttpServletRequest httpServletRequest) {
+		for (InfoDisplayRequestAttributesContributor
+				infoDisplayRequestAttributesContributor :
+					_infoDisplayRequestAttributesContributors) {
+
+			infoDisplayRequestAttributesContributor.addAttributes(
+				httpServletRequest);
+		}
 	}
 
 	@Override
@@ -260,7 +259,7 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 	}
 
 	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.content)",
+		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.display.page)",
 		unbind = "-"
 	)
 	protected void setServletContext(ServletContext servletContext) {
@@ -293,13 +292,11 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 		PermissionChecker permissionChecker, Layout layout) {
 
 		try {
-			if (_layoutPermission.contains(
+			if (LayoutPermissionUtil.contains(
 					permissionChecker, layout, ActionKeys.UPDATE) ||
-				_layoutPermission.contains(
+				LayoutPermissionUtil.contains(
 					permissionChecker, layout,
-					ActionKeys.UPDATE_LAYOUT_CONTENT) ||
-				_modelResourcePermission.contains(
-					permissionChecker, layout.getPlid(), ActionKeys.UPDATE)) {
+					ActionKeys.UPDATE_LAYOUT_CONTENT)) {
 
 				return true;
 			}
@@ -313,17 +310,16 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 		return false;
 	}
 
-	private static final String _EDIT_LAYOUT_PAGE =
-		"/layout/edit_layout/content.jsp";
+	private static final String _EDIT_PAGE = "/layout/edit/display_page.jsp";
 
 	private static final String _URL =
 		"${liferay:mainPath}/portal/layout?p_l_id=${liferay:plid}" +
 			"&p_v_l_s_g_id=${liferay:pvlsgid}";
 
-	private static final String _VIEW_PAGE = "/layout/view/content.jsp";
+	private static final String _VIEW_PAGE = "/layout/view/display_page.jsp";
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		ContentLayoutTypeController.class);
+		DisplayPageLayoutTypeController.class);
 
 	@Reference
 	private FragmentCollectionContributorTracker
@@ -337,6 +333,10 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Reference
 	private InfoDisplayContributorTracker _infoDisplayContributorTracker;
+
+	@Reference
+	private volatile List<InfoDisplayRequestAttributesContributor>
+		_infoDisplayRequestAttributesContributors;
 
 	@Reference
 	private InfoItemRendererTracker _infoItemRendererTracker;
@@ -353,12 +353,6 @@ public class ContentLayoutTypeController extends BaseLayoutTypeControllerImpl {
 	@Reference
 	private LayoutPageTemplateEntryLocalService
 		_layoutPageTemplateEntryLocalService;
-
-	@Reference
-	private LayoutPermission _layoutPermission;
-
-	@Reference
-	private LayoutContentModelResourcePermission _modelResourcePermission;
 
 	@Reference
 	private Portal _portal;
