+++ modules/apps/layout/layout-type-controller/layout-type-controller-link-to-page/src/main/java/com/liferay/layout/type/controller/link/to/page/internal/layout/type/controller/LinkToPageLayoutTypeController.java
@@ -12,9 +12,12 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.panel.internal.controller;
+package com.liferay.layout.type.controller.link.to.page.internal.layout.type.controller;
 
+import com.liferay.item.selector.ItemSelector;
 import com.liferay.layout.type.controller.BaseLayoutTypeControllerImpl;
+import com.liferay.layout.type.controller.link.to.page.internal.constants.LinkToPageLayoutTypeControllerWebKeys;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutConstants;
@@ -31,13 +34,20 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Eudaldo Alonso
+ * @author Pavel Savinov
  */
 @Component(
-	immediate = true, property = "layout.type=" + LayoutConstants.TYPE_PANEL,
+	immediate = true,
+	property = "layout.type=" + LayoutConstants.TYPE_LINK_TO_LAYOUT,
 	service = LayoutTypeController.class
 )
-public class PanelLayoutTypeController extends BaseLayoutTypeControllerImpl {
+public class LinkToPageLayoutTypeController
+	extends BaseLayoutTypeControllerImpl {
+
+	@Override
+	public String getType() {
+		return LayoutConstants.TYPE_LINK_TO_LAYOUT;
+	}
 
 	@Override
 	public String getURL() {
@@ -50,6 +60,8 @@ public class PanelLayoutTypeController extends BaseLayoutTypeControllerImpl {
 			HttpServletResponse httpServletResponse, Layout layout)
 		throws Exception {
 
+		httpServletRequest.setAttribute(
+			LinkToPageLayoutTypeControllerWebKeys.ITEM_SELECTOR, _itemSelector);
 		httpServletRequest.setAttribute(WebKeys.SEL_LAYOUT, layout);
 
 		return super.includeEditContent(
@@ -58,7 +70,7 @@ public class PanelLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public boolean isFirstPageable() {
-		return true;
+		return false;
 	}
 
 	@Override
@@ -68,12 +80,17 @@ public class PanelLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public boolean isSitemapable() {
-		return true;
+		return false;
 	}
 
 	@Override
 	public boolean isURLFriendliable() {
 		return true;
+	}
+
+	@Override
+	public boolean isWorkflowEnabled() {
+		return false;
 	}
 
 	@Override
@@ -92,23 +109,25 @@ public class PanelLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	protected String getViewPage() {
-		return _VIEW_PAGE;
+		return StringPool.BLANK;
 	}
 
 	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.panel)",
+		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.link.to.page)",
 		unbind = "-"
 	)
 	protected void setServletContext(ServletContext servletContext) {
 		this.servletContext = servletContext;
 	}
 
-	private static final String _EDIT_PAGE = "/layout/edit/panel.jsp";
+	private static final String _EDIT_PAGE = "/layout/edit/link_to_layout.jsp";
 
 	private static final String _URL =
-		"${liferay:mainPath}/portal/layout?p_l_id=${liferay:plid}" +
-			"&p_v_l_s_g_id=${liferay:pvlsgid}";
+		"${liferay:mainPath}/portal/layout?p_v_l_s_g_id=${liferay:pvlsgid}&" +
+			"groupId=${liferay:groupId}&privateLayout=${privateLayout}&" +
+				"layoutId=${linkToLayoutId}";
 
-	private static final String _VIEW_PAGE = "/layout/view/panel.jsp";
+	@Reference
+	private ItemSelector _itemSelector;
 
 }
