+++ modules/apps/layout/layout-type-controller/layout-type-controller-node/src/main/java/com/liferay/layout/type/controller/node/internal/layout/type/controller/NodeLayoutTypeController.java
@@ -12,33 +12,31 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.url.internal.controller;
+package com.liferay.layout.type.controller.node.internal.layout.type.controller;
 
 import com.liferay.layout.type.controller.BaseLayoutTypeControllerImpl;
-import com.liferay.petra.string.StringPool;
+import com.liferay.layout.type.controller.node.internal.constants.NodeLayoutTypeControllerConstants;
+import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
-import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutConstants;
 import com.liferay.portal.kernel.model.LayoutTypeController;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.taglib.servlet.PipingServletResponse;
 
 import javax.servlet.ServletContext;
 import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Eudaldo Alonso
+ * @author Juergen Kappler
  */
 @Component(
-	immediate = true, property = "layout.type=" + LayoutConstants.TYPE_URL,
+	immediate = true,
+	property = "layout.type=" + NodeLayoutTypeControllerConstants.LAYOUT_TYPE_NODE,
 	service = LayoutTypeController.class
 )
-public class URLLayoutTypeController extends BaseLayoutTypeControllerImpl {
+public class NodeLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public String getURL() {
@@ -46,15 +44,8 @@ public class URLLayoutTypeController extends BaseLayoutTypeControllerImpl {
 	}
 
 	@Override
-	public String includeEditContent(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse, Layout layout)
-		throws Exception {
-
-		httpServletRequest.setAttribute(WebKeys.SEL_LAYOUT, layout);
-
-		return super.includeEditContent(
-			httpServletRequest, httpServletResponse, layout);
+	public boolean isBrowsable() {
+		return false;
 	}
 
 	@Override
@@ -64,7 +55,7 @@ public class URLLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public boolean isParentable() {
-		return false;
+		return true;
 	}
 
 	@Override
@@ -74,7 +65,7 @@ public class URLLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	public boolean isURLFriendliable() {
-		return true;
+		return false;
 	}
 
 	@Override
@@ -98,19 +89,24 @@ public class URLLayoutTypeController extends BaseLayoutTypeControllerImpl {
 
 	@Override
 	protected String getViewPage() {
-		return StringPool.BLANK;
+		return _VIEW_PAGE;
 	}
 
 	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.url)",
+		target = "(osgi.web.symbolicname=com.liferay.layout.type.controller.node)",
 		unbind = "-"
 	)
 	protected void setServletContext(ServletContext servletContext) {
 		this.servletContext = servletContext;
 	}
 
-	private static final String _EDIT_PAGE = "/layout/edit/url.jsp";
+	private static final String _EDIT_PAGE = "/layout/edit/node.jsp";
 
-	private static final String _URL = "${url}";
+	private static final String _URL = StringBundler.concat(
+		"${liferay:mainPath}/portal/layout?p_v_l_s_g_id=${liferay:pvlsgid}&",
+		"groupId=${liferay:groupId}&privateLayout=${liferay:privateLayout}&",
+		"layoutId=${liferay:layoutId}");
+
+	private static final String _VIEW_PAGE = "/layout/view/node.jsp";
 
 }
